requirements.txt
pandas
scikit-learn==0.23.2
scipy==1.7.1
mlflow==2.8.1
azureml-mlflow==1.51.0
azureml-inference-server-http
azureml-core==1.49.0
cloudpickle==1.6.0
joblib